feat: Authentication & API Docs
- Use Scribe for API Docs
- Use Sanctum for authentication

+++ app/Http/Controllers/API/GenreApiController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\API;
 
 use App\Http\Controllers\API\ApiBaseController;
 use App\Http\Controllers\Controller;
+use App\Http\Requests\PaginationAPIRequest;
 use App\Http\Requests\StoreGenreAPIRequest;
 use App\Http\Requests\UpdateGenreAPIRequest;
 use App\Models\Genre;
@@ -11,19 +12,35 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Support\Js;
+use Knuckles\Scribe\Attributes\Response;
+use Knuckles\Scribe\Attributes\ResponseField;
+use Knuckles\Scribe\Attributes\UrlParam;
 use Psy\Util\Json;
-use function PHPUnit\Framework\isNull;
 
 class GenreApiController extends ApiBaseController
 {
-    /**
-     * Display a listing of the resource.
-     */
-    public function index(): JsonResponse
-    {
-        $genres = Genre::all();
 
-        if (!is_null($genres) && $genres->count() > 0) {
+    /**
+     * api/genres
+     *
+     * 200 List of all genres.
+     */
+    #[UrlParam("id", "int", "The author's ID.", required: true, example: 7)]
+    #[Response('"authors": [{"id": 7,"given_name": "Kevin","family_name": "Potts","is_company": 0,"created_at": "2022-09-10T14:45:22.000000Z","updated_at": "2022-09-10T14:45:22.000000Z"}]', 200, "Retrieved successfully.")]
+    #[ResponseField("status", "Success or failure indicator.")]
+    #[ResponseField("message", "Accompanying message for the status result.")]
+    #[ResponseField("authors", "The author details.")]
+    public function index(PaginationAPIRequest $request): JsonResponse
+    {
+//        if (isset($request['all']) && $request['all']) {
+//            $genres = Genre::all();
+//        } else {
+//            $genres = Genre::paginate($request['per_page']);
+//        }
+        $allGenres = $request['all'] === true || $request['all'] === 1;
+        $genres = $allGenres ? Genre::all() :Genre::paginate($request['per_page']);
+
+        if (!is_null($genres) && $genres && $genres->count() > 0) {
             return $this->sendResponse(
                 $genres,
                 "Retrieved successfully.",
@@ -42,7 +59,7 @@ class GenreApiController extends ApiBaseController
 
         $results = Genre::create($validated);
 
-        if (!is_null($results) && $results->count() > 0) {
+        if (!is_null($results) && $genre && $results->count() > 0) {
             return $this->sendResponse(
                 $results,
                 "Created genre successfully.",
@@ -60,7 +77,7 @@ class GenreApiController extends ApiBaseController
 
         $genre = Genre::whereUuid($uuid)->first();
 
-        if (isset($genre) && $genre->count() > 0) {
+        if (!is_null($genre) && $genre && $genre->count() > 0) {
             return $this->sendResponse(
                 $genre,
                 "Retrieved successfully.",
@@ -77,7 +94,7 @@ class GenreApiController extends ApiBaseController
     {
         $genre = Genre::whereUuid($uuid)->first();
 
-        if (isset($genre) && !is_null($genre)) {
+        if (!is_null($genre) && $genre && $genre !== false) {
 
             $validated = $request->validated();
 
@@ -103,7 +120,7 @@ class GenreApiController extends ApiBaseController
         $genre = Genre::whereUuid($uuid)->first();
         $results = $genre;
 
-        if (!is_null($genre) && $genre->count() > 0) {
+        if (!is_null($genre) && $genre && $genre->count() > 0) {
 
             $deleted = $genre->delete();
